+++ packages/uui-card-user/lib/uui-card-user.element.ts
@@ -13,6 +13,7 @@ import { ifDefined } from 'lit/directives/if-defined.js';
  *  @description - Card component for displaying a user node.
  *  @slot - slot for the default content area
  *  @slot tag - slot for the tag with support for `<uui-tag>` elements
+ *  @slot avatar - slot for the avatar with support for the `<uui-avatar>` element
  *  @slot actions - slot for the actions with support for the `<uui-action-bar>` element
  */
 @defineElement('uui-card-user')