+++ packages/uui-color-picker/lib/uui-color-picker.element.ts
@@ -74,7 +74,7 @@ export class UUIColorPickerElement extends LabelMixin('label', LitElement) {
   @state() private hue = 0;
   @state() private saturation = 0;
   @state() private lightness = 0;
-  @state() private alpha = 0;
+  @state() private alpha = 100;
   @state() private _colord: Colord = colord('hsl(0, 0%, 0%)');
 
   /**
@@ -361,7 +361,7 @@ export class UUIColorPickerElement extends LabelMixin('label', LitElement) {
     if (colorString === this.value) return;
 
     if (!colorString) {
-      this.alpha = 0;
+      this.alpha = 100;
       this.inputValue = '';
       this._value = colorString;
 
@@ -379,7 +379,7 @@ export class UUIColorPickerElement extends LabelMixin('label', LitElement) {
     this.hue = h;
     this.saturation = s;
     this.lightness = l;
-    this.alpha = this.opacity ? a * 100 : 100;
+    this.alpha = this.opacity ? a * 100 : 100; // Convert to 0-100 range, and set alpha to 100 if opacity is disabled
 
     const hslaColor = colorString as HslaColor;
 
@@ -395,7 +395,6 @@ export class UUIColorPickerElement extends LabelMixin('label', LitElement) {
     this.dispatchEvent(
       new UUIColorPickerChangeEvent(UUIColorPickerChangeEvent.CHANGE),
     );
-
     return true;
   }
 
@@ -531,7 +530,7 @@ export class UUIColorPickerElement extends LabelMixin('label', LitElement) {
   }
 
   private _renderSwatches() {
-    if (!this.swatches) return nothing;
+    if (!this.swatches?.length) return nothing;
     return html`<uui-color-swatches
       id="swatches"
       class="color-picker__swatches"
@@ -586,6 +585,7 @@ export class UUIColorPickerElement extends LabelMixin('label', LitElement) {
         --uui-look-outline-border: #ddd;
         --uui-look-outline-border-hover: #aaa;
         font-size: 0.8rem;
+        color: var(--uui-color-text);
         display: block;
         width: var(--uui-color-picker-width, 280px);
       }
@@ -597,9 +597,9 @@ export class UUIColorPickerElement extends LabelMixin('label', LitElement) {
       }
       .color-picker {
         width: 100%;
-        background-color: #fff;
+        background-color: var(--uui-color-surface);
         user-select: none;
-        border: solid 1px #e4e4e7;
+        border: solid 1px var(--uui-color-border);
       }
       .color-picker__user-input {
         display: flex;
@@ -641,7 +641,6 @@ export class UUIColorPickerElement extends LabelMixin('label', LitElement) {
         width: 100%;
         height: 100%;
         border-radius: inherit;
-        box-shadow: inset 0 0 0 1px rgba(0, 0, 0, 0.2);
         /* We use a custom property in lieu of currentColor because of https://bugs.webkit.org/show_bug.cgi?id=216780 */
         background-color: var(--preview-color);
       }
@@ -651,6 +650,7 @@ export class UUIColorPickerElement extends LabelMixin('label', LitElement) {
       }
 
       .color-picker__transparent-bg {
+        border: 1px solid var(--uui-color-border);
         background-image: linear-gradient(
             45deg,
             var(--uui-palette-grey) 25%,
@@ -705,7 +705,7 @@ export class UUIColorPickerElement extends LabelMixin('label', LitElement) {
       }
 
       uui-color-swatches {
-        border-top: solid 1px #d4d4d8;
+        border-top: solid 1px var(--uui-color-border);
         padding: 0.75rem;
       }